refactor(buildrequest): extract retry policy into retry_policy.go
Move ResolveRetryPolicy, EncodeRetryPolicy, and RetryConfigToPolicy from
orchestrator.go into a new retry_policy.go (~125 LOC). Same package,
verbatim move with no behaviour change. Regenerates bamlutils/embed.go
to include the new file (worker-plugin compilation reads from the
embedded FS).

Part of #199 item 4.
